feat: add renderAsList filter
