+++ .app/core/query.js
@@ -1,4 +1,5 @@
 const operators = require("./query-operators");
+const ValueParser = require("./value-parser");
 
 module.exports = class QueryRunner {
   constructor(items, query) {
@@ -129,21 +130,5 @@ class QuerySorter {
 
   compareValues(a, b) {
     return a < b ? -1 : a > b ? 1 : 0;
-  }
-}
-
-class ValueParser {
-  static getValueByPath(item, path) {
-    if (!path) return item;
-
-    const parts = path.split(".");
-    let value = item;
-
-    for (let part of parts) {
-      if (!(part in value)) return undefined;
-      value = value[part];
-    }
-
-    return value;
   }
 }
