+++ .app/package.json
@@ -21,6 +21,7 @@
     "markdown-it-footnote": "^3.0.3",
     "markdown-it-task-checkbox": "^1.0.6",
     "markdown-it-wikilinks": "^1.2.0",
+    "nanohtml": "^1.10.0",
     "node-html-parser": "^6.1.4",
     "npm-run-all": "^4.1.5",
     "parcel": "^2.8.2",
